Move Attributes form to separate component
